+++ src/internal.hpp
@@ -1562,9 +1562,9 @@ struct Internal {
 
   // Internal functions to enable explicit search limits.
   //
-  void limit_terminate (int);
-  void limit_decisions (int);     // Force decision limit.
-  void limit_conflicts (int);     // Force conflict limit.
+  void limit_terminate (int64_t);
+  void limit_decisions (int64_t);     // Force decision limit.
+  void limit_conflicts (int64_t);     // Force conflict limit.
   void limit_preprocessing (int); // Enable 'n' preprocessing rounds.
   void limit_local_search (int);  // Enable 'n' local search rounds.
   void limit_ticks (int64_t);     // Force ticks limit.
@@ -1572,7 +1572,7 @@ struct Internal {
   // External versions can access limits by 'name'.
   //
   static bool is_valid_limit (const char *name);
-  bool limit (const char *name, int); // 'true' if 'name' valid
+  bool limit (const char *name, int64_t); // 'true' if 'name' valid
 
   // Set all the CDCL search limits and increments for scheduling
   // inprocessing, restarts, clause database reductions, etc.
